tests(ImageSelector): relocate preview related tests
this was necessary to circumvent the react-redux `connect` method which prevented me from easily mock `getMediaItem` but I also think tests related to the preview component should live in their own file
